DM-179 updated naming of DisplayAvatar, removed unncessary useeffect and fetching
